Add contract validation during LocalModel creation
